Face errors.
I have faced a lot of issues, because of version difference of Redux.
and also It takes a lot of time to solve the issues by learning
concepts.
I realized that understanding the code is much more important than just
following lecture so that i can solve issues on myside when it happens,

so I change my study strategy.